requirements.txt
Django==4.0.6
numpy==1.23.0
opencv_python==4.6.0.66
Pillow==9.2.0
